api: make Call = Call17
Closes #235
