--- const_call_17.go
@@ -1,8 +1,0 @@
-//go:build go_tarantool_call_17
-// +build go_tarantool_call_17
-
-package tarantool
-
-const (
-	CallRequestCode = Call17RequestCode
-)
