+++ request.go
@@ -218,9 +218,7 @@ func (conn *Connection) Upsert(space interface{}, tuple, ops interface{}) (resp 
 }
 
 // Call calls registered Tarantool function.
-// It uses request code for Tarantool >= 1.7 if go-tarantool
-// was build with go_tarantool_call_17 tag.
-// Otherwise, uses request code for Tarantool 1.6.
+// It uses request code for Tarantool >= 1.7, result is an array.
 //
 // It is equal to conn.CallAsync(functionName, args).Get().
 func (conn *Connection) Call(functionName string, args interface{}) (resp *Response, err error) {
@@ -228,7 +226,7 @@ func (conn *Connection) Call(functionName string, args interface{}) (resp *Respo
 }
 
 // Call16 calls registered Tarantool function.
-// It uses request code for Tarantool 1.6, so result is converted to array of arrays
+// It uses request code for Tarantool 1.6, result is an array of arrays.
 // Deprecated since Tarantool 1.7.2.
 //
 // It is equal to conn.Call16Async(functionName, args).Get().
@@ -237,8 +235,7 @@ func (conn *Connection) Call16(functionName string, args interface{}) (resp *Res
 }
 
 // Call17 calls registered Tarantool function.
-// It uses request code for Tarantool >= 1.7, so result is not converted
-// (though, keep in mind, result is always array)
+// It uses request code for Tarantool >= 1.7, result is an array.
 //
 // It is equal to conn.Call17Async(functionName, args).Get().
 func (conn *Connection) Call17(functionName string, args interface{}) (resp *Response, err error) {
@@ -329,9 +326,7 @@ func (conn *Connection) UpdateTyped(space, index interface{}, key, ops interface
 }
 
 // CallTyped calls registered function.
-// It uses request code for Tarantool >= 1.7 if go-tarantool
-// was build with go_tarantool_call_17 tag.
-// Otherwise, uses request code for Tarantool 1.6.
+// It uses request code for Tarantool >= 1.7, result is an array.
 //
 // It is equal to conn.Call16Async(functionName, args).GetTyped(&result).
 func (conn *Connection) CallTyped(functionName string, args interface{}, result interface{}) (err error) {
@@ -339,7 +334,7 @@ func (conn *Connection) CallTyped(functionName string, args interface{}, result 
 }
 
 // Call16Typed calls registered function.
-// It uses request code for Tarantool 1.6, so result is converted to array of arrays
+// It uses request code for Tarantool 1.6, result is an array of arrays.
 // Deprecated since Tarantool 1.7.2.
 //
 // It is equal to conn.Call16Async(functionName, args).GetTyped(&result).
@@ -348,8 +343,7 @@ func (conn *Connection) Call16Typed(functionName string, args interface{}, resul
 }
 
 // Call17Typed calls registered function.
-// It uses request code for Tarantool >= 1.7, so result is not converted
-// (though, keep in mind, result is always array)
+// It uses request code for Tarantool >= 1.7, result is an array.
 //
 // It is equal to conn.Call17Async(functionName, args).GetTyped(&result).
 func (conn *Connection) Call17Typed(functionName string, args interface{}, result interface{}) (err error) {
@@ -422,16 +416,14 @@ func (conn *Connection) UpsertAsync(space interface{}, tuple interface{}, ops in
 }
 
 // CallAsync sends a call to registered Tarantool function and returns Future.
-// It uses request code for Tarantool >= 1.7 if go-tarantool
-// was build with go_tarantool_call_17 tag.
-// Otherwise, uses request code for Tarantool 1.6.
+// It uses request code for Tarantool >= 1.7, so future's result is an array.
 func (conn *Connection) CallAsync(functionName string, args interface{}) *Future {
 	req := NewCallRequest(functionName).Args(args)
 	return conn.Do(req)
 }
 
 // Call16Async sends a call to registered Tarantool function and returns Future.
-// It uses request code for Tarantool 1.6, so future's result is always array of arrays.
+// It uses request code for Tarantool 1.6, so future's result is an array of arrays.
 // Deprecated since Tarantool 1.7.2.
 func (conn *Connection) Call16Async(functionName string, args interface{}) *Future {
 	req := NewCall16Request(functionName).Args(args)
@@ -439,8 +431,7 @@ func (conn *Connection) Call16Async(functionName string, args interface{}) *Futu
 }
 
 // Call17Async sends a call to registered Tarantool function and returns Future.
-// It uses request code for Tarantool >= 1.7, so future's result will not be converted
-// (though, keep in mind, result is always array)
+// It uses request code for Tarantool >= 1.7, so future's result is an array.
 func (conn *Connection) Call17Async(functionName string, args interface{}) *Future {
 	req := NewCall17Request(functionName).Args(args)
 	return conn.Do(req)
@@ -1115,9 +1106,8 @@ type CallRequest struct {
 	args     interface{}
 }
 
-// NewCallRequest return a new empty CallRequest. It uses request code for
-// Tarantool >= 1.7 if go-tarantool was build with go_tarantool_call_17 tag.
-// Otherwise, uses request code for Tarantool 1.6.
+// NewCallRequest returns a new empty CallRequest. It uses request code for
+// Tarantool >= 1.7.
 func NewCallRequest(function string) *CallRequest {
 	req := new(CallRequest)
 	req.requestCode = CallRequestCode
